Windows fixes (#21)
- Fixes matrix dir-file separation by using the right separator
- Stop canonicalizing paths on windows.
https://doc.rust-lang.org/std/fs/fn.canonicalize.html#platform-specific-behavior
- Let glob algorithms operate on linux style separators only

This last point is a bit meh in terms of perf. But at least it works.
Just calls `replace` on the paths and patterns, but of course this is
often not really needed (especially on Linux).
So let me know if you deem it acceptable or whether you want me to
change it.